requirements.txt
python-bitcoinlib>=0.11.0;<0.12.0
click
